// lib/pte/types.ts
export type SpeakingType =
  | 'read_aloud'
  | 'repeat_sentence'
  | 'describe_image'
  | 'retell_lecture'
  | 'answer_short_question'
  | 'summarize_group_discussion'
  | 'respond_to_a_situation'